@@ -8,14 +8,14 @@
     "start": "next start",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
-    "lint2": "run -T eslint --max-warnings 0 .",
-    "lint:fix": "run -T eslint --fix .",
+    "lint": "eslint --max-warnings 0 .",
+    "lint:fix": "eslint --fix .",
     "test": "vitest",
     "test:ci": "vitest run",
     "test:coverage": "vitest run --coverage"
   },
   "dependencies": {
-    "@dotkom/ui": "workspace:^",
+    "@dotkomonline/ui": "workspace:^",
     "@radix-ui/colors": "^0.1.8",
     "@radix-ui/react-avatar": "^1.0.0",
     "@radix-ui/react-checkbox": "^1.0.0",
@@ -26,9 +26,9 @@
     "@sanity/client": "^2.23.2",
     "@stitches/react": "^1.2.8",
     "axios": "^0.21.4",
-    "eslint-config-next": "^12.2.3",
     "luxon": "^2.5.0",
-    "next": "^12.3.0",
+    "next": "13.0.0",
+    "next-auth": "^4.10.3",
     "next-compose-plugins": "^2.2.1",
     "next-transpile-modules": "^9.0.0",
     "prettier": "^2.7.1",
@@ -39,29 +39,29 @@
     "swr": "^1.3.0"
   },
   "devDependencies": {
-    "@dotkom/eslint-config-ow": "workspace:*",
+    "@dotkomonline/config": "workspace:*",
     "@storybook/addon-actions": "^6.5.9",
     "@storybook/addon-essentials": "^6.5.9",
     "@storybook/addon-knobs": "^6.4.0",
     "@storybook/addon-links": "^6.5.9",
     "@storybook/preset-scss": "^1.0.3",
     "@storybook/react": "^6.5.9",
+    "@types/eslint": "^8",
     "@types/jsdom": "^20",
     "@types/luxon": "^2.4.0",
     "@types/node": "^16.11.47",
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
     "@vitejs/plugin-react": "^2.1.0",
     "@vitest/coverage-c8": "^0.23.2",
+    "autoprefixer": "^10.4.13",
     "chromatic": "^6.7.3",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-prettier": "^3.4.1",
+    "eslint": "^8.26.0",
+    "eslint-config-next": "^13.0.0",
     "husky": "^4.3.8",
     "jsdom": "^20.0.0",
     "lint-staged": "^10.5.4",
+    "postcss": "^8.4.18",
     "tsconfig-paths-webpack-plugin": "^4.0.0",
     "typescript": "^4.7.4",
     "vite": "^3.1.1",
