@@ -1,5 +1,5 @@
+import { User as PrismaUser } from "@dotkomonline/db"
 import { z } from "zod"
-import { User as PrismaUser } from "@dotkom/db"
 
 const userSchema = z.object({
   id: z.string().uuid(),
