@@ -1,14 +1,15 @@
+import { Company as PrismaCompany } from "@dotkomonline/db"
 import { z } from "zod"
-import { Company as PrismaCompany } from "@dotkom/db"
 
 const companySchema = z.object({
   id: z.string().uuid(),
   name: z.string().max(50),
-  description: z.string().optional(),
+  description: z.string(),
   phone: z.string().optional(),
-  email: z.string().optional(),
-  website: z.string().optional(),
+  email: z.string(),
+  website: z.string(),
   location: z.string().optional(),
+  type: z.enum(["Consulting", "Research", "Development", "Other"]).optional(),
 })
 
 export type Company = z.infer<typeof companySchema>
