@@ -1,5 +1,5 @@
 import { initTRPC } from "@trpc/server"
 
-export const t = initTRPC.create();
+import { Context } from "./index.js"
 
-export const appRouter = t.router({})
+export const t = initTRPC.context<Context>().create()
