@@ -1,8 +1,9 @@
 /// <reference types="vitest" />
+import tsconfigPaths from "vite-tsconfig-paths"
 import { defineConfig } from "vitest/config"
 
 export default defineConfig({
-  plugins: [],
+  plugins: [tsconfigPaths()],
   test: {
     globals: true,
     environment: "node",
