@@ -1,15 +1,15 @@
 /* eslint-disable */
-const withPlugins = require("next-compose-plugins")
-const transpileModules = require("next-transpile-modules")(["@dotkom/ui"]);
+const withTM = require("next-transpile-modules")(["@dotkomonline/ui"])
 
 /**
  * @type {import('next').NextConfig}
  */
-const config = {
+const nextConfig = {
   swcMinify: true,
   images: {
     domains: ["cdn.sanity.io"],
   },
+  reactStrictMode: true,
 }
 
-module.exports = withPlugins([transpileModules], config)
+module.exports = withTM(nextConfig)
