@@ -1,6 +1,7 @@
+import { css, styled } from "@dotkomonline/ui"
 import { mauve, violet } from "@radix-ui/colors"
-import { css, styled } from "@theme"
 import { FC } from "react"
+
 import { NavigationMenuLink } from "."
 import { DesktopProps } from "./DesktopProps"
 
