@@ -1,13 +1,13 @@
+import Content from "./Content"
 import ContentList from "./ContentList"
 import ContentListItem from "./ContentListItem"
 import DebugLink from "./DebugLink"
-import NavigationMenuViewport from "./NavigationMenuViewport"
-import Link from "./Link"
-import Menu from "./Menu"
-import List from "./List"
 import IndicatorWithArrow from "./IndicatorWithArrow"
+import Link from "./Link"
+import List from "./List"
+import Menu from "./Menu"
+import NavigationMenuViewport from "./NavigationMenuViewport"
 import TriggerWithCaret from "./Trigger"
-import Content from "./Content"
 
 export const DesktopList = ContentList
 export const DesktopNavbarContainer = Menu
