@@ -4,7 +4,7 @@ const useWindow = () => {
   const [color, setColor] = useState("transparent")
   const [shadow, setShadow] = useState("0 0.1px 0.2px #0d0d0d")
 
-  const changeNavbarColor = () => setColor("#fff")
+  const changeNavbarColor = () => setColor("rgba(255,255,255, 0.8)")
   const revertNavbarColor = () => setColor("transparent")
 
   const changeShadow = () => setShadow("0 0.1px 1px #0d0d0d")
