@@ -1,4 +1,4 @@
-import { styled } from "@theme"
+import { styled } from "@dotkomonline/ui"
 
 export const Box = styled("div")
 export const Flex = styled(Box, { display: "flex" })
