@@ -1,11 +1,13 @@
+import { CSS, css, styled } from "@dotkomonline/ui"
+import { Button } from "@dotkomonline/ui"
+import { BlockContentProps } from "@sanity/block-content-to-react"
+import { FC } from "react"
+
 import PortableText from "@components/molecules/PortableText"
 import { Box } from "@components/primitives"
-import { BlockContentProps } from "@sanity/block-content-to-react"
-import { CSS, css, styled } from "@theme"
-import { FC } from "react"
+
 import CompanyInterestProcess from "./CompanyInterestProcess"
 import OurProducts from "./OurProducts"
-import { Button } from "@dotkom/ui"
 
 export type Content = BlockContentProps["blocks"]
 
