@@ -1,4 +1,4 @@
-import { getCssText } from "@dotkom/ui"
+import { getCssText } from "@dotkomonline/ui"
 import NextDocument, { Html, Head, Main, NextScript, DocumentContext, DocumentInitialProps } from "next/document"
 
 export default class Document extends NextDocument {
@@ -12,7 +12,7 @@ export default class Document extends NextDocument {
       <Html lang="en">
         <Head>
           <link
-            href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;600;700&display=swap"
+            href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;700&display=swap"
             rel="stylesheet"
           />
           <style id="stitches" dangerouslySetInnerHTML={{ __html: getCssText() }} />
