@@ -1,6 +1,13 @@
-import { globalCss } from "@stitches/react"
+import { globalCss } from "@dotkomonline/ui"
 
 export const globalStyles = globalCss({
+  html: {
+    fontSize: "14px",
+  },
+  "html, body, #__next": {
+    width: "100%",
+    height: "100%",
+  },
   body: {
     fontFamily: "$body",
     margin: 0,
