@@ -1,7 +1,7 @@
 /// <reference types="vitest" />
-import { defineConfig } from "vitest/config"
 import react from "@vitejs/plugin-react"
 import tsconfigPaths from "vite-tsconfig-paths"
+import { defineConfig } from "vitest/config"
 
 export default defineConfig({
   plugins: [react(), tsconfigPaths()],
