@@ -1,5 +1,6 @@
-import { useState } from "react"
 import { ComponentStory } from "@storybook/react"
+import { useState } from "react"
+
 import { TextInput } from "./TextInput"
 
 export default {
@@ -18,7 +19,7 @@ const Template: ComponentStory<typeof TextInput> = (args) => {
 export const Default = Template.bind({})
 
 export const BrandWithAsterisk = Template.bind({})
-BrandWithAsterisk.args = { variant: "brand", withAsterisk: true }
+BrandWithAsterisk.args = { withAsterisk: true }
 
 export const Placeholder = Template.bind({})
 Placeholder.args = { placeholder: "Chad Thunder" }
@@ -27,7 +28,7 @@ export const Error = Template.bind({})
 Error.args = { error: true, value: "Invalid text" }
 
 export const ErrorWithMessage = Template.bind({})
-ErrorWithMessage.args = { error: "Too short", placeholder: "Shrimp gang" }
+ErrorWithMessage.args = { error: "Too short", value: "Shrimp gang" }
 
 export const Disabled = Template.bind({})
 Disabled.args = { disabled: true, value: "Do not touch my pringles" }
